+++ code/src/TransactionFailEvent.cpp
@@ -17,7 +17,11 @@ TransactionFailEvent::TransactionFailEvent() :
 {
 }
 
-TransactionFailEvent::TransactionFailEvent(ActorId ship_id, ActorId port_id, double ship_gold, double ship_rum, double port_rum)
+TransactionFailEvent::TransactionFailEvent(ActorId ship_id,
+					   ActorId port_id,
+					   unsigned int ship_gold,
+					   unsigned int ship_rum,
+					   unsigned int port_rum)
   : EventInterface(),
     ship_id(ship_id),
     port_id(port_id),
